cast

+++ include/glaze/binary/header.hpp
@@ -146,7 +146,7 @@ namespace glz::detail
    {
       uint8_t header;
       std::memcpy(&header, &(*it), 1);
-      const uint8_t config = get_bits<2>(header);
+      const uint8_t config = uint8_t(get_bits<2>(header));
 
       switch (config) {
       case 0:
